--- a/src/afterfact.rs
+++ b/src/afterfact.rs
@@ -1806,6 +1806,7 @@ mod tests {
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv.csv").to_path_buf()),
             multiline: false,
+            remove_duplicate_data: false,
         });
         let dummy_config = Some(Config {
             action: Some(dummy_action),
@@ -2099,6 +2100,7 @@ mod tests {
             geo_ip: None,
             output: Some(Path::new("./test_emit_csv_multiline.csv").to_path_buf()),
             multiline: true,
+            remove_duplicate_data: false,
         });
         let dummy_config = Some(Config {
             action: Some(dummy_action),