--- a/src/afterfact.rs
+++ b/src/afterfact.rs
@@ -323,6 +323,7 @@ fn emit_csv<W: std::io::Write>(
             .unwrap_or(&0) as usize
     };
     let mut author_list_cache: HashMap<CompactString, Nested<String>> = HashMap::new();
+    let mut prev_message: HashMap<CompactString, Profile> = HashMap::new();
     for (message_idx, time) in MESSAGEKEYS
         .lock()
         .unwrap()
@@ -331,7 +332,6 @@ fn emit_csv<W: std::io::Write>(
         .enumerate()
     {
         // remove duplicate dataのための前レコード分の情報を保持する変数
-        let mut prev_message: HashMap<CompactString, Profile> = HashMap::new();
         let multi = message::MESSAGES.get(time).unwrap();
         let (_, detect_infos) = multi.pair();
         timestamps[message_idx] = _get_timestamp(output_option, time);