--- a/src/detections/rule/condition_parser.rs
+++ b/src/detections/rule/condition_parser.rs
@@ -598,6 +598,7 @@ mod tests {
                 geo_ip: None,
                 output: None,
                 multiline: false,
+                remove_duplicate_data: false,
             })),
             debug: false,
         }))