--- a/src/detections/rule/count.rs
+++ b/src/detections/rule/count.rs
@@ -622,6 +622,7 @@ mod tests {
                 geo_ip: None,
                 output: None,
                 multiline: false,
+                remove_duplicate_data: false,
             })),
             debug: false,
         }))