## What Changed

- Added `-R, --remove-duplicate-data` option in `csv-timeline` command, for remove duplicate `Details` field (`%Details%`, `%AllFieldInfo%`, `%ExtraFieldInfo%`) output. 

I would appreciate it if you could review when you have time.