@YamatoSecurity 2023-06-22
@hitenkoku Thanks so much for the PR! @fukusuket Thanks for checking!
I ran some tests comparing the main 2.7.0-dev branch and this PR with and without the `-R` option.

This PR does a good job at reducing filesize of about 10% (10GB to 9GB), however, it seems to miss events but detects one more unique detection.

|                   | Main 2.7.0-dev | This PR without -R | This PR with -R |
|-------------------|----------------|--------------------|-----------------|
| Elasped Time      | 02:26:16.238   | 02:27:08.033       | 02:27:01.602    |
| Memory Usage      | 65.4 GB        | 49.4 GB            | 49.4 GB         |
| File Size         | 13.9 GB        | 10.0 GB            | 9.0 GB          |
| Total Detections  | 25,142,009     | 18,728,720         | 18,728,720      |
| Unique Detections | 370            | 371                | 371             |

Details:
```
Main 2.7.0-dev branch:
.\hayabusa-2.7.0-dev-main.exe csv-timeline -d D:\logs\608-eventlogs\ -u -U -D  -o main.csv --debug

First Timestamp: 2010-11-21 03:58:31.093 +00:00
Last Timestamp: 2020-02-25 17:17:00.835 +00:00

Events with hits / Total events: 22,749,640 / 73,049,133 (Data reduction: 50,299,493 events (68.86%))

Total | Unique detections: 25,142,009 | 370
Total | Unique critical detections: 95 (0.00%) | 7 (1.89%)
Total | Unique high detections: 86,152 (0.34%) | 95 (25.68%)
Total | Unique medium detections: 3,173,286 (12.62%) | 132 (35.68%)
Total | Unique low detections: 1,418,409 (5.64%) | 74 (20.00%)
Total | Unique informational detections: 20,464,067 (81.39%) | 62 (16.76%)

Saved file: main.csv (13.9 GB)

Elapsed time: 02:26:16.238
Errors were generated. Please check ./logs/errorlog-20230623_050444.log for details.

Rule Parse Processing Time: 00:00:02.369
Analysis Processing Time: 02:18:48.753
Output Processing Time: 00:07:25.071

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:    72.0 GiB    72.0 GiB     0          72.0 GiB
 committed:    65.5 GiB   339.3 GiB   274.7 GiB    64.6 GiB
     reset:     1.0 GiB
    purged:   152.9 GiB
   touched:    64.2 KiB    35.8 MiB   756.9 GiB  -756.9 GiB                          ok
  segments:    44         573         560          13                                not all freed!
-abandoned:     0           0           0           0                                ok
   -cached:     0           0           0           0                                ok
     pages:     0           0          10.5 Mi    -10.5 Mi                           ok
-abandoned:     0           0           0           0                                ok
 -extended:     0
 -noretire:     0
     mmaps:     0
   commits:     2.5 Mi
    resets:    35
    purges:   304.9 Ki
   threads:    16          16           0          16                                not all freed!
  searches:     0.0 avg
numa nodes:     1
   elapsed:  8776.247 s
   process: user: 59178.203 s, system: 469.656 s, faults: 68466755, rss: 56.0 GiB, commit: 65.4 GiB
```

This PR without `-R`:
```
.\hayabusa-remove-dups.exe csv-timeline -d D:\logs\608-eventlogs\ -u -U -D  -o no-dups.csv --debug

First Timestamp: 2010-11-21 03:58:31.093 +00:00
Last Timestamp: 2020-02-25 17:17:00.835 +00:00

Events with hits / Total events: 16,352,779 / 73,049,133 (Data reduction: 56,696,354 events (77.61%))

Total | Unique detections: 18,728,720 | 371
Total | Unique critical detections: 95 (0.00%) | 7 (1.89%)
Total | Unique high detections: 86,279 (0.46%) | 95 (25.61%)
Total | Unique medium detections: 3,173,358 (16.94%) | 134 (36.12%)
Total | Unique low detections: 1,418,347 (7.57%) | 73 (19.68%)
Total | Unique informational detections: 14,050,641 (75.02%) | 62 (16.71%)

Saved file: no-dups.csv (10.0 GB)

Elapsed time: 02:27:08.033
Errors were generated. Please check ./logs/errorlog-20230622_224055.log for details.

Rule Parse Processing Time: 00:00:05.018
Analysis Processing Time: 02:22:54.612
Output Processing Time: 00:04:08.169

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:    52.0 GiB    52.0 GiB     0          52.0 GiB
 committed:    98.0 GiB   570.4 GiB   522.2 GiB    48.1 GiB
     reset:     1.0 GiB
    purged:   300.7 GiB
   touched:    64.2 KiB   270.8 MiB   757.4 GiB  -757.1 GiB                          ok
  segments:    32           4.2 Ki      4.2 Ki     12                                not all freed!
-abandoned:     0           0           0           0                                ok
   -cached:     0           0           0           0                                ok
     pages:     0           0          10.5 Mi    -10.5 Mi                           ok
-abandoned:     0           0           0           0                                ok
 -extended:     0
 -noretire:     0
     mmaps:     0
   commits:     3.7 Mi
    resets:    40
    purges:   296.7 Ki
   threads:    16          16           0          16                                not all freed!
  searches:     0.0 avg
numa nodes:     1
   elapsed:  8830.956 s
   process: user: 60674.171 s, system: 439.875 s, faults: 94348959, rss: 48.3 GiB, commit: 49.4 GiB
```

This PR with `-R`:
```
.\hayabusa-remove-dups.exe csv-timeline -d D:\logs\608-eventlogs\ -u -U -D -R -o dups.csv --debug

Results Summary:
First Timestamp: 2010-11-21 03:58:31.093 +00:00
Last Timestamp: 2020-02-25 17:17:00.835 +00:00

Events with hits / Total events: 16,352,779 / 73,049,133 (Data reduction: 56,696,354 events (77.61%))

Total | Unique detections: 18,728,720 | 371
Total | Unique critical detections: 95 (0.00%) | 7 (1.89%)
Total | Unique high detections: 86,279 (0.46%) | 95 (25.61%)
Total | Unique medium detections: 3,173,358 (16.94%) | 134 (36.12%)
Total | Unique low detections: 1,418,347 (7.57%) | 73 (19.68%)
Total | Unique informational detections: 14,050,641 (75.02%) | 62 (16.71%)

Saved file: dups.csv (9.0 GB)

Elapsed time: 02:27:01.602
Errors were generated. Please check ./logs/errorlog-20230622_185156.log for details.

Rule Parse Processing Time: 00:00:04.072
Analysis Processing Time: 02:22:59.460
Output Processing Time: 00:03:58.056

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:    52.0 GiB    52.0 GiB     0          52.0 GiB
 committed:   100.9 GiB   575.2 GiB   524.4 GiB    50.8 GiB
     reset:     1.1 GiB
    purged:   303.2 GiB
   touched:    64.2 KiB   270.5 MiB   757.3 GiB  -757.1 GiB                          ok
  segments:    32           4.2 Ki      4.2 Ki     12                                not all freed!
-abandoned:     0           0           0           0                                ok
   -cached:     0           0           0           0                                ok
     pages:     0           0          10.5 Mi    -10.5 Mi                           ok
-abandoned:     0           0           0           0                                ok
 -extended:     0
 -noretire:     0
     mmaps:     0
   commits:     3.8 Mi
    resets:    39
    purges:   297.1 Ki
   threads:    16          16           0          16                                not all freed!
  searches:     0.0 avg
numa nodes:     1
   elapsed:  8821.611 s
   process: user: 60742.281 s, system: 427.531 s, faults: 94995342, rss: 48.3 GiB, commit: 49.4 GiB
```


Besides missing events, there is some difference in the fields:
Main:
`"2018-08-11 15:47:15.929 +00:00","xxx","Sec",4634,"info",10977375,"Logoff","User: HealthMailbox250cf44 ¦ LID: 0x1021972ec","LogonType: 3 ¦ TargetDomainName: shieldbase ¦ TargetUserSid: S-1-5-21-3445421715-2530590580-3149308974-1163"`

This PR:
`"2018-08-11 15:47:15.929 +00:00","xxx","Sec",4634,"info",10977375,"Logoff","User: HealthMailbox250cf44 ¦ LID: 0x1021972ec ¦ Type: 3","TargetDomainName: shieldbase ¦ TargetUserSid: S-1-5-21-3445421715-2530590580-3149308974-1163"`

However, the %Details% for EID 4634 is `'User: %TargetUserName% ¦ LID: %TargetLogonId% ¦ Type: %LogonType%'` so this PR seems correct using the `Type` field and the main branch seems incorrect by using the `LogonType` field and putting it in the `%ExtraField%` column. This PR puts it in the right `%Details%` column.

I will take some more benchmarks with the updated PR.